Tighten wait condition [5.1.x] (#259)
I made a pass on the rest of tests that have not moved to TopologyTestDriver, and they are all a bit cumbersome to convert, because 1) either rely on Avro SR embedded with CLUSTER, 2) or reply on existing XXXService which is not easily integrated with TopologyTestDriver.

So what I did here is the following:

Generally, many of the flaky tests is not because Streams client takes too long to rebalance, but because broker cluster is taking too longer to start-up, and create / delete topics, and hence unable to handle the JoinGroup Request yet. For example, most recent failures:
https://jenkins.confluent.io/job/confluentinc/job/kafka-streams-examples/job/5.1.x/274/testReport/junit/io.confluent.examples.streams.microservices/ValidationsAggregatorServiceTest/shouldAggregateRuleSuccesses/

https://jenkins.confluent.io/job/confluentinc/job/kafka-streams-examples/job/5.2.x/182/testReport/junit/io.confluent.examples.streams.microservices/OrdersServiceTest/shouldGetOrderByIdWhenOnDifferentHost/

Falls into this scenario. So what I did is to tighten the wait condition on create / delete topics, which are usually done at the startup of test, by checking both the ZK path as well as broker's local caches. This also makes sure that broker has completed startup and completed the topic operation. Note previously, we do not check after deleting the topic, and hence when starting the streams app the broker may still be busy deleting the topic and hence cause time out issue.

Another minor fix, is that for IQ tests, a lot of the time binding fails because of the random host. I've removed this from the logic but always use localhost instead. I think this does not lose generality of the test coverage.